revert `layout` component changes
